requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
Jinja2
python-multipart
requests
httptools
kafka-python
tqdm
minio
openpyxl
xlsxwriter
msoffcrypto-tool
pyxlsb